+++ 3_PhaseTransition/GameByFunctions/main_GameByFunctions.cpp
@@ -13,7 +13,7 @@
 SDL_Window* g_window;
 SDL_Renderer* g_renderer;
 bool g_flag_running;
-Uint32 g_last_time_ms;
+Uint64 g_last_time_ms;
 MIX_Mixer* g_mixer;
 
 
@@ -43,7 +43,7 @@ int main(int argc, char* argv[])
 
 	while ( g_flag_running )
 	{
-		Uint32 cur_time_ms = SDL_GetTicks();
+		Uint64 cur_time_ms = SDL_GetTicks();
 		if ( cur_time_ms-g_last_time_ms < 33 )
 			continue;
