+++ 4_Projectile/FreeFall/G2W.cpp
@@ -1,24 +1,24 @@
 #include "G2W.h"
 
-double G2W_Scale = 300.0;
-double W2G_Scale = 1.f / 300.0;
+float G2W_Scale = 300.0f;
+float W2G_Scale = 1.f / 300.0f;
 
-int G2W_X(double g_x)
+float G2W_X(float g_x)
 {
-	return (int)(g_x * G2W_Scale + 500);
+	return g_x * G2W_Scale + 500.0f;
 }
 
-int G2W_Y(double g_y)
+float G2W_Y(float g_y)
 {
-	return (int)(-1*g_y * G2W_Scale + 650);
+	return -1.0f*g_y * G2W_Scale + 650.0f;
 }
 
-double W2G_X(int w_x)
+float W2G_X(float w_x)
 {
 	return (w_x - 500.f) * W2G_Scale;
 }
 
-double W2G_Y(int w_y)
+float W2G_Y(float w_y)
 {
 	return (w_y - 650.f) * -1 * W2G_Scale;
 }
